+++ src/main/java/controller/Controller.java
@@ -1,0 +1,10 @@
+package controller;
+
+import http.request.HttpRequest;
+import http.response.HttpResponse;
+
+public interface Controller {
+
+    void service(HttpRequest request, HttpResponse response);
+
+}
