+++ src/main/java/controller/CreateUserController.java
@@ -1,0 +1,19 @@
+package controller;
+
+import db.DataBase;
+import http.request.HttpRequest;
+import http.response.HttpResponse;
+import model.User;
+
+public class CreateUserController implements Controller{
+
+    @Override
+    public void service(HttpRequest request, HttpResponse response) {
+        User user = new User(request.getParameter("userId"),
+            request.getParameter("password"),
+            request.getParameter("name"),
+            request.getParameter("email"));
+        DataBase.addUser(user);
+        response.sendRedirect("/index.html");
+    }
+}
