+++ src/main/java/webserver/RequestHandler.java
@@ -1,26 +1,19 @@
 package webserver;
 
-import db.DataBase;
+import controller.Controller;
 import http.request.HttpRequest;
 import http.response.HttpResponse;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.Socket;
-import java.util.Collection;
-import java.util.Map;
-import login.LoginService;
-import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import util.HttpRequestUtils;
 
 public class RequestHandler extends Thread {
     private static final Logger log = LoggerFactory.getLogger(RequestHandler.class);
 
     private final Socket connection;
-    private final LoginService loginService = LoginService.getInstance();
-
 
     public RequestHandler(Socket connectionSocket) {
         this.connection = connectionSocket;
@@ -34,77 +27,17 @@ public class RequestHandler extends Thread {
         try (InputStream in = connection.getInputStream(); OutputStream out = connection.getOutputStream()) {
             HttpRequest request = new HttpRequest(in);
             HttpResponse response = new HttpResponse(out);
-            String path = getDefaultPath(request.getPath());
 
-            // create 요청
-            if ("/user/create".equals(path)) {
-                createUser(request, response);
-            } else if ("/user/login".equals(path)) {
-                login(request, response);
-            } else if ("/user/list".equals(path)) {
-                listUser(request, response);
-            } else {
+            Controller controller = RequestMapping.getController(request.getPath());
+            if (controller == null) {
+                String path = getDefaultPath(request.getPath());
                 response.forwardBody(path);
+            } else {
+                controller.service(request, response);
             }
         } catch (IOException e) {
             log.error(e.getMessage());
         }
-    }
-
-    private void listUser(HttpRequest request, HttpResponse response) {
-        if (!isLogin(request.getHeader("Cookie"))) {
-            response.sendRedirect("/user/login.html");
-            return;
-        }
-
-        Collection<User> users = DataBase.findAll();
-        StringBuilder sb = new StringBuilder();
-        sb.append("<table border='1'>");
-        for (User user : users) {
-            sb.append("<tr>");
-            sb.append("<tr>" + user.getUserId() + "</td>");
-            sb.append("<tr>" + user.getName() + "</td>");
-            sb.append("<tr>" + user.getEmail() + "</td>");
-            sb.append("</tr>");
-        }
-        response.forwardBody(sb.toString());
-    }
-
-    private static void login(HttpRequest request, HttpResponse response) {
-        User findUser = DataBase.findUserById(request.getParameter("userId"));
-        if (findUser != null) {
-            // 존재하는 회원
-            if (findUser.login(request.getParameter("password"))) {
-                // 비밀번호 맞으면 로그인 성공
-                response.addHeader("Set-Cookie", "logined=true; Path=/");
-                response.sendRedirect("/index.html");
-            } else {
-                // 비밀 번호 틀리면 로그인 실패
-                response.sendRedirect("/user/login_failed.html");
-            }
-        } else {
-            response.sendRedirect("/user/login_failed.html");
-        }
-    }
-
-    private static void createUser(HttpRequest request, HttpResponse response) {
-        User user = new User(request.getParameter("userId"),
-            request.getParameter("password"),
-            request.getParameter("name"),
-            request.getParameter("email"));
-
-        log.debug("user: {}", user);
-        DataBase.addUser(user);
-        response.sendRedirect("/index.html");
-    }
-
-    private boolean isLogin(String cookie) {
-        Map<String, String> cookies = HttpRequestUtils.parseCookies(cookie);
-        String value = cookies.get("logined");
-        if (value == null) {
-            return false;
-        }
-        return Boolean.parseBoolean(value);
     }
 
     private String getDefaultPath(String path) {
