+++ src/main/java/webserver/RequestMapping.java
@@ -1,0 +1,24 @@
+package webserver;
+
+import controller.Controller;
+import controller.CreateUserController;
+import controller.ListUserController;
+import controller.LoginController;
+import java.util.HashMap;
+import java.util.Map;
+
+public class RequestMapping {
+
+    private static Map<String, Controller> controllers = new HashMap<>();
+
+    static {
+        controllers.put("/user/create", new CreateUserController());
+        controllers.put("/user/login", new LoginController());
+        controllers.put("/user/list", new ListUserController());
+    }
+
+    public static Controller getController(String requestUrl) {
+        return controllers.get(requestUrl);
+    }
+
+}
